Added uri, method and http version checks in processRequest()
Add: Added elifs checks in processRequest() to check for invalid headers
and return 400 in case of missing method or missing uri. Returns 505 in
case of unsupported http version.

Issue: Error response is wrong if the method is empty. Uri appears to
contain the http version instead? Need to beef up the constructor
probably.