Various tweaks
Tweak: Reverted port listening to 8085 back to 8080 because I cannot
fucking stand my browser cache defaulting to 8080 and jumpscaring me
with the error ever.

Change: Swap makeErrorResponse() call to makeResponse() for 204 no
content in Response.cpp

Change: HTTPRequest constructor checks for GET POST and DELETE methods
in its _method member, setting it to INVALID in case of diff.
processRequest() then checks for INVALID and returns
makeErrorReponse(405...) if INVALID is found.